Move analysis to root for now
Try to find a better place for it, but it needs to go for now to avoid circular imports
