+++ spacy/analysis.py
@@ -4,8 +4,8 @@ from __future__ import unicode_literals
 from collections import OrderedDict
 from wasabi import Printer
 
-from ..tokens import Doc, Token, Span
-from ..errors import user_warning
+from .tokens import Doc, Token, Span
+from .errors import user_warning
 
 
 def analyze_pipes(pipeline, name, pipe, index, warn=True):
